Add files via upload
This is the shuttle data of interest, combined with the weather data. In the main Jupyter notebook I had uploaded, I converted merged_df to df at a certain point to make it easier on myself, which is why it shows as such near the bottom of the notebook, but feel free to approach as you wish.